+++ variables.tf
@@ -335,10 +335,6 @@ variable "core_services_config" {
     prometheus = optional(object({
       remote_write = optional(any, [])
     }), {})
-    storage = optional(object({
-      file = optional(bool, false)
-      blob = optional(bool, false)
-    }), {})
   })
   nullable = false
 }